+++ lib/rivium_push.dart
@@ -5,5 +5,6 @@ export 'src/rivium_push_config.dart';
 export 'src/rivium_push_message.dart';
 export 'src/rivium_push_error.dart';
 export 'src/inapp_message.dart';
+export 'src/inapp/rivium_push_inapp_overlay.dart';
 export 'src/inbox_message.dart';
 export 'src/ab_test.dart';